+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.13.0]
+## [1.13.0] - 2016-01-17
 ### Added
 - Python 3.5 has been added as a tested environment for this package.